requirements.txt
python-binance==0.7.11
python-coinmarketcap==0.2